removed SUPPRESS_EMPTY_FILE_WARNING from AKU-debugger; fixed gcc build warning in MOAIHttpServer

+++ src/moaiext-server/MOAIHttpServer.cpp
@@ -131,7 +131,7 @@ int MOAIHttpServer::HandleRequest ( mg_connection* connection, mg_request_info* 
 				"%s",
 				statusCode,
 				contentType,
-				contentLength,
+				( int )contentLength,
 				content
 			);
